+++ src/mvc/Model/Network/Antenna/GeoPosition/GeoPosition.hpp
@@ -5,17 +5,18 @@
 
 class GeoPosition {
     public :
-        GeoPosition(sf::Vector2<float> lambert);
+        GeoPosition(sf::Vector2<double> lambert);
         ~GeoPosition();
-        sf::Vector2<float> lambert() const;
-        sf::Vector2<float> coords() const;
-        void setLambert(const sf::Vector2<float>& lambert);
-        void setCoords(const sf::Vector2<float>& coords);
+        //replace returned type to Lambert struct
+        sf::Vector2<double> lambert() const;
+        sf::Vector2<double> coords() const;
+        void setLambert(const sf::Vector2<double>& lambert);
+        void setCoords(const sf::Vector2<double>& coords);
         void L2C();
         void C2L();
     private :
-        sf::Vector2<float> m_lambert;
-        sf::Vector2<float> m_coords;
+        sf::Vector2<double> m_lambert;
+        sf::Vector2<double> m_coords;
 };
